Moved the tests to a new test file.
